refactor: new card stack component
Move the "stack of cards" behavior to its own WC as this will be reused later for Leitner System boxes, and the
responsibility of dc-repetition was multiple without any performance benefits. In order to show custom events, and to
keep the scope of this refactoring small, dc-repetition kept the responsibility of the actions on the stack (ok/nok),
but this may be behind this new component responsibility instead.